pyproject.toml
[tool.poetry]
name = "update-stars"
version = "0.1.0"
description = ""
authors = ["Stefane Fermigier <sf@fermigier.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PyGithub = "^1.54"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
update-stars = 'update_stars:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
